Decouple weapon properties in style of other Item components.
