+++ src/main/java/org/dungeon/creatures/AttackAlgorithm.java
@@ -117,7 +117,7 @@ class AttackAlgorithm {
     // Check that there is a weapon and that it is not broken.
     if (weapon != null && !weapon.isBroken()) {
       if (weapon.rollForHit()) {
-        hitDamage = weapon.getDamage() + attacker.getAttack();
+        hitDamage = weapon.getWeaponComponent().getDamage() + attacker.getAttack();
         weapon.decrementIntegrityByHit();
         weaponBroke = weapon.isBroken();
       } else {
@@ -156,7 +156,7 @@ class AttackAlgorithm {
       // Check that there is a weapon and that it is not broken.
       if (weapon != null && !weapon.isBroken()) {
         if (weapon.rollForHit()) {
-          hitDamage = weapon.getDamage() + attacker.getAttack();
+          hitDamage = weapon.getWeaponComponent().getDamage() + attacker.getAttack();
           criticalHit = Engine.roll(HERO_CRITICAL_CHANCE);
           weapon.decrementIntegrityByHit();
           weaponBroke = weapon.isBroken();
